Update: Main Menu + .csv Fix
This first update added a functional menu to the system, as well as fixed an error with the .csv file. In previous versions, the .csv file would not load due to numerous errors in its formatting, leading me and Mr. Scott to finding a solution online, where the system would ignore erroneous entries allowing the system work.