Update: Basic Menu Completion
This update finishes the basic menu. The menu consists of an option to see the original dataset, search for more specific datasets and quit the program. In previous tests, the 'search for ore specific' option did not have further options stopping the user from continuing further. In this update, I have added further options within the option, however those options are yet to do anything and will be added later on.